hardening http transport manager: make proxy part of key
