+++ src/main/extension/extension_install.cpp
@@ -1,7 +1,7 @@
 #include "duckdb/common/exception/http_exception.hpp"
 #include "duckdb/common/gzip_file_system.hpp"
-#include "duckdb/common/http_transport_manager.hpp"
-#include "duckdb/common/http_util.hpp"
+#include "duckdb/main/http/http_transport_manager.hpp"
+#include "duckdb/main/http/http_util.hpp"
 #include "duckdb/common/local_file_system.hpp"
 #include "duckdb/main/database_file_opener.hpp"
 #include "duckdb/common/serializer/binary_serializer.hpp"
